refactor: remove superficial sqlschema.Operation interface
Each dialect has to type-switch the operation before building a query for it.
Since the migrator knows the concrete type of each operation, they are free
to provide FQN in any form. Using schema.FQN field from the start simplifies
the data structure later.

Empty inteface is better that a superficial one.